+++ R/data.R
@@ -14,6 +14,44 @@
 #' @source example data
 "data_bio2rarify"
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# "data_fish_MBSS" ####
+#' @title Fish data, MBSS
+#'
+#' @description A dataset with example fish taxa data for metric calculation.
+#' @format{ A data frame with 1694 observations on the following 27 variables.
+#' \describe{
+#'   \item{\code{SAMPLEID}}{a character vector}
+#'   \item{\code{TAXAID}}{a character vector}
+#'   \item{\code{N_TAXA}}{a numeric vector}
+#'   \item{\code{TYPE}}{a character vector}
+#'   \item{\code{TOLER}}{a character vector}
+#'   \item{\code{NATIVE}}{a character vector}
+#'   \item{\code{TROPHIC}}{a character vector}
+#'   \item{\code{SILT}}{a character vector}
+#'   \item{\code{INDEX_REGION}}{a character vector}
+#'   \item{\code{SAMP_LENGTH_M}}{a numeric vector}
+#'   \item{\code{SAMP_WIDTH_M}}{a numeric vector}
+#'   \item{\code{SAMP_BIOMASS}}{a numeric vector}
+#'   \item{\code{INDEX_NAME}}{a character vector}
+#'   \item{\code{DA_MI2}}{a numeric vector}
+#'   \item{\code{N_ANOMALIES}}{a numeric vector}
+#'   \item{\code{FAMILY}}{a character vector}
+#'   \item{\code{GENUS}}{a character vector}
+#'   \item{\code{BCG_ATTR}}{a character vector}
+#'   \item{\code{THERMAL_INDICATOR}}{a character vector}
+#'   \item{\code{ELEVATION_ATTR}}{a character vector}
+#'   \item{\code{GRADIENT_ATTR}}{a character vector}
+#'   \item{\code{WSAREA_ATTR}}{a character vector}
+#'   \item{\code{REPRODUCTION}}{a character vector}
+#'   \item{\code{HABITAT}}{a character vector}
+#'   \item{\code{CONNECTIVITY}}{a logical vector}
+#'   \item{\code{SCC}}{a logical vector}
+#'   \item{\code{HYBRID}}{a logical vector}
+#' }
+#' }
+#' @source example data
+"data_fish_MBSS"
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # "data_fish_MA" ####
 #' @title Estuary taxa data
 #'