+++ tests/testthat/test_metric_names.R
@@ -1,7 +1,7 @@
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # Bugs----
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-## metval, Names, bugs, Function, xlNames ####
+## metval, names, bugs, Function, xlNames ####
 test_that("metric.values, names, bugs, Function, xlNames", {
   # Packages
   #library(readxl) # part of BioMonTools
@@ -133,7 +133,7 @@ test_that("metric.values, names, bugs, Function, xlScoring", {
   metnam_xlScoring[!(metnam_xlScoring %in% metnam_fun)]
 
   # test
-  testthat::expect_equal(metnam_len, metnam_match)  # fails due to structure
+  testthat::expect_equal(metnam_len, metnam_match)
 })## Test ~ metric.values, names, bugs, Function, Names ~ END
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -197,46 +197,125 @@ test_that("metric.scores, index, number metrics", {
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # Fish ----
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-## metval, Names, fish, Function, xlNames ####
+## metval, names, fish, Function, xlNames ####
 # Need fish data
-test_that("metric.values, names, algae, Function, xlNames", {
-  # Packages
-  # #library(readxl) # part of BioMonTools
-  #
-  # # Data
-  # fn_metnam_xlNames   <- file.path(system.file(package="BioMonTools")
-  #                                  , "extdata"
-  #                                  , "MetricNames.xlsx")
-  # # Import
-  # df_metnam_xlNames   <- readxl::read_excel(fn_metnam_xlNames
-  #                                           , sheet="MetricMetadata"
-  #                                           , skip = 4)
-  # df_metnam_xlNames <- as.data.frame(df_metnam_xlNames)
-  # # Metric Names
-  # metnam_xlNames <- df_metnam_xlNames[df_metnam_xlNames[, "Community"]=="algae"
-  #                                     , "METRIC_NAME"
-  #                                     , drop = TRUE]
-  #
-  # # Algae data
-  # df_diatoms <- BioMonTools::data_diatom_mmi_dev
-  #
-  # # Function
-  # df_metval <- BioMonTools::metric.values(df_diatoms
-  #                                         , "algae"
-  #                                         , boo.Shiny = TRUE)
-  # metnam_fun <- colnames(df_metval)[-c(1:3)] # remove first 3 columns
-  #
-  # # Check
-  # metnam_len <- length(metnam_fun)
-  # metnam_match <- sum(metnam_fun %in% metnam_xlNames)
-  #
-  # # Show non-matches
-  # metnam_fun[!(metnam_fun %in% metnam_xlNames)]
-  #
-  # # test
-  # testthat::expect_equal(metnam_len, metnam_match)
+test_that("metric.values, names, fish, Function, xlNames", {
+  #Packages
+  #library(readxl) # part of BioMonTools
+
+  # Data
+  fn_metnam_xlNames   <- file.path(system.file(package="BioMonTools")
+                                   , "extdata"
+                                   , "MetricNames.xlsx")
+  # Import
+  df_metnam_xlNames   <- readxl::read_excel(fn_metnam_xlNames
+                                            , sheet="MetricMetadata"
+                                            , skip = 4)
+  df_metnam_xlNames <- as.data.frame(df_metnam_xlNames)
+  # Metric Names
+  metnam_xlNames <- df_metnam_xlNames[df_metnam_xlNames[, "Community"]=="fish"
+                                      , "METRIC_NAME"
+                                      , drop = TRUE]
+
+  # fish data
+  df_fish <- BioMonTools::data_fish_MBSS
+
+  # Function
+  df_metval <- BioMonTools::metric.values(df_fish
+                                          , "fish"
+                                          , boo.Shiny = TRUE)
+  metnam_fun <- colnames(df_metval)[-c(1:5)] # remove first few columns
+
+  # Check
+  metnam_len <- length(metnam_fun)
+  metnam_match <- sum(metnam_fun %in% metnam_xlNames)
+
+  # Show non-matches
+  metnam_fun[!(metnam_fun %in% metnam_xlNames)]
+
+  # test
+  testthat::expect_equal(metnam_len, metnam_match)
 
 })## Test ~ metric names, fish, Function, Names ~ END
+
+## metval, names, fish, xlScoring, xlNames ####
+test_that("metric.values, names, fish, xlNames, xlScoring", {
+  # Packages
+  #library(readxl) # part of BioMonTools
+
+  # Data
+  fn_metnam_xlNames   <- file.path(system.file(package="BioMonTools")
+                                   , "extdata"
+                                   , "MetricNames.xlsx")
+  fn_metnam_xlScoring <- file.path(system.file(package="BioMonTools")
+                                   , "extdata"
+                                   , "MetricScoring.xlsx")
+  # Import
+  df_metnam_xlNames   <- readxl::read_excel(fn_metnam_xlNames
+                                            , sheet="MetricMetadata"
+                                            , skip = 4)
+  df_metnam_xlScoring <- readxl::read_excel(fn_metnam_xlScoring
+                                            , sheet="metric.scoring")
+  # Convert to data frames
+  df_metnam_xlNames <- as.data.frame(df_metnam_xlNames)
+  df_metnam_xlScoring <- as.data.frame(df_metnam_xlScoring)
+
+  # Names
+  metnam_xlNames <- df_metnam_xlNames[df_metnam_xlNames[, "Community"]=="fish"
+                                      , "METRIC_NAME", drop = TRUE]
+  metnam_xlScoring <- unique(df_metnam_xlScoring[df_metnam_xlScoring[
+    , "Community"]=="fish", "METRIC_NAME", drop = TRUE])
+
+  # Check
+  metnam_xlScoring_match <- sum(metnam_xlScoring %in% metnam_xlNames)
+  metnam_xlScoring_len <- length(metnam_xlScoring)
+
+  # Show non-matches
+  metnam_xlScoring[!(metnam_xlScoring %in% metnam_xlNames)]
+
+  # test
+  testthat::expect_equal(metnam_xlScoring_match, metnam_xlScoring_len)
+})## Test - metric.values, names, Excel, Scoring ~ END
+
+## metval, names, fish, xlScoring, Function  ####
+test_that("metric.values, names, fish, Function, xlScoring", {
+  # Packages
+  #library(readxl) # part of BioMonTools
+
+  # Data
+  fn_metnam_xlScoring <- file.path(system.file(package="BioMonTools")
+                                   , "extdata"
+                                   , "MetricScoring.xlsx")
+  # Import
+  df_metnam_xlScoring <- readxl::read_excel(fn_metnam_xlScoring
+                                            , sheet="metric.scoring")
+  df_metnam_xlScoring <- as.data.frame(df_metnam_xlScoring)
+
+  # Metric Names
+  metnam_xlScoring <- unique(df_metnam_xlScoring[df_metnam_xlScoring[
+    , "Community"]=="fish", "METRIC_NAME", drop = TRUE])
+
+  # Benthic Data
+  df_fish <- BioMonTools::data_fish_MBSS
+  #df_benthos$SUBCLASS <- NA
+
+  # Function
+  df_metval <- BioMonTools::metric.values(df_fish
+                                          , "fish"
+                                          , boo.Shiny = TRUE)
+  metnam_fun <- colnames(df_metval)[-c(1:5)] # remove first few columns
+
+  # Check
+  metnam_len <- length(metnam_xlScoring)
+  metnam_match <- sum(metnam_xlScoring %in% metnam_fun)
+
+  # Show non-matches
+  metnam_xlScoring[!(metnam_xlScoring %in% metnam_fun)]
+
+  # test
+  testthat::expect_equal(metnam_len, metnam_match)
+})## Test ~ metric.values, names, bugs, Function, Names ~ END
+
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # Algae ----